# core/vm: improved EVM run loop & instruction calling