**Review #3348 first**

This PR implements a refactored EVM and EVM run loop.

### Simplified run loop

The run loop, which previously contained custom opcode executes have been removed and has been simplified to a few checks; including:

* memory expansion
* gas calculation & metering
* instruction execution

Each operation consists out of 4 elements:

* Execution function
* gas cost function
* stack validation function
* memory size function

The **execution function** implements the operation's runtime behaviour, the **gas cost function** implements the operation gas costs function and greatly depends on the memory and stack, the **stack validation function** validates the stack and makes sure that enough items can be popped off and pushed on and the **memory size function** calculates the memory required for the operation and returns it.

### Optional gas metering

This PR also allows the EVM to go unmetered. This is helpful for offline operations such as contract `call`'s.