@GitCop 2016-12-01
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: fa061147d800bfb92a83373cd6901ba02c3935f3
  * Commits must be prefixed with the package(s) they modify

 * Commit: aa7342e7d12e0b9dae366ff2daa9a164507d54b4
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com