@GitCop 2016-12-02
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: 121d1d34ec00ed217487c1ade4fbffdb57694892
  * Commits must be prefixed with the package(s) they modify

 * Commit: ae381031c6a1fa98a2ddb49cb49a69d33977cf6b
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com