@GitCop 2016-12-02
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: 121d1d34ec00ed217487c1ade4fbffdb57694892
  * Commits must be prefixed with the package(s) they modify

 * Commit: b457d147020335be48c2c482b490c92d1f4d3d58
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com