@obscuren 2016-12-02
Preliminary benchmarks have given the following results:

**Old vs new (metered):**

benchmark                         old ns/op     new ns/op     delta
BenchmarkVmFibonacci16Tests-4     37668307      21323060      -43.39%

**Old vs new (unmetered):**

benchmark                         old ns/op     new ns/op     delta
BenchmarkVmFibonacci16Tests-4     37668307      14204094      -62.29%