@GitCop 2016-12-15
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: d3752868ece61e9e18f85fe022799c10590ca3bd
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com